Update init database

+++ swp/src/main/java/org/swp/repository/IBookingRepository.java
@@ -3,5 +3,5 @@ package org.swp.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.swp.entity.Booking;
 
-public interface IBookingRepository extends JpaRepository<Booking, Integer> {
+public interface IBookingRepository extends JpaRepository<Booking, Long> {
 }
